refactor swarm to not being dependent on kubernetes, but the only implementation right now is kubernetes
