+++ swarm/swarm.go
@@ -1,8 +1,7 @@
 package swarm
 
 import (
-	"bytes"
-	"encoding/gob"
+	"errors"
 	"fmt"
 	"net"
 	"time"
@@ -11,15 +10,27 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-// NodeState represents the current state of a cluster node known by
-// this instance.
-type NodeState int
+type swarmType int
 
 const (
-	Initial NodeState = iota
-	Connected
-	Disconnected
+	swarmKubernetes swarmType = iota
+	swarmUnknown
 )
+
+func (st swarmType) String() string {
+	switch st {
+	case swarmKubernetes:
+		return "kubernetes Swarm"
+	}
+	return "unkwown Swarm"
+}
+
+func getSwarmType(o Options) swarmType {
+	if o.KubernetesOptions.KubernetesInCluster && o.KubernetesOptions.KubernetesAPIBaseURL != "" {
+		return swarmKubernetes
+	}
+	return swarmUnknown
+}
 
 const (
 	// DefaultMaxMessageBuffer is the default maximum size of the
@@ -33,110 +44,28 @@ const (
 	DefaultLeaveTimeout = time.Duration(5 * time.Second)
 )
 
-// NodeInfo is a value object tat contains information about swarm
-// cluster nodes, that is required to access member nodes.
-type NodeInfo struct {
-	Name string
-	Addr net.IP
-	Port int
-}
+var (
+	ErrUnknownSwarm = errors.New("unknown swarm type")
+)
 
-func (ni *NodeInfo) String() string {
-	return fmt.Sprintf("NodeInfo{%s, %s, %d}", ni.Name, ni.Addr, ni.Port)
-}
-
-func mapNodesToAddresses(n []*NodeInfo) []string {
-	var s []string
-	for i := range n {
-		s = append(s, fmt.Sprintf("%v:%d", n[i].Addr, n[i].Port))
-	}
-
-	return s
-}
-
-// Self can return itself as NodeInfo
-type Self interface {
-	Node() *NodeInfo
-}
-
-// EntryPoint knows its peers of nodes which contains itself
-type EntryPoint interface {
-	Nodes() []*NodeInfo
-}
-
-// knownEntryPoint is the Kubernetes based implementation of the
-// interfaces Self and Entrypoint. It stores an initial list of peers
-// to connect to at the start.
-type knownEntryPoint struct {
-	self  *NodeInfo
-	nodes []*NodeInfo
-}
-
-// newKnownEntryPoint returns a new knownEntryPoint that knows all
-// initial peers and itself. If it can not get a list of peers it will
-// fail fast.
-func newKnownEntryPoint(o Options) *knownEntryPoint {
-	nic := NewnodeInfoClient(o)
-	nodes, err := nic.GetNodeInfo()
-	if err != nil {
-		log.Fatalf("SWARM: Failed to get nodeinfo: %v", err)
-	}
-	self := getSelf(nodes)
-	log.Infof("SWARM: Join swarm self=%s, nodes=%v", self, nodes)
-	return &knownEntryPoint{self: self, nodes: nodes}
-}
-
-// Node return its self
-func (e *knownEntryPoint) Node() *NodeInfo {
-	if e == nil {
-		return nil
-	}
-	return e.self
-}
-
-func (e *knownEntryPoint) Nodes() []*NodeInfo {
-	if e == nil {
-		return nil
-	}
-	return e.nodes
-}
-
-func getSelf(nodes []*NodeInfo) *NodeInfo {
-	addrs, err := net.InterfaceAddrs()
-	if err != nil {
-		log.Fatalf("SWARM: Failed to get addr: %v", err)
-	}
-
-	for _, ni := range nodes {
-		for _, addr := range addrs {
-			ip, _, err := net.ParseCIDR(addr.String())
-			if err != nil {
-				log.Errorf("SWARM: could not parse cidr: %v", err)
-				continue
-			}
-			if ip.Equal(ni.Addr) {
-				return ni
-			}
-		}
-	}
-	return nil
-}
-
-// Options for swarm objects.
-type Options struct {
-	// leaky, expected to be buffered, or errors are lost
-	Errors chan<- error
-
-	MaxMessageBuffer int
-
-	LeaveTimeout time.Duration
-
+// KubernetesOptions specific swarm options
+type KubernetesOptions struct {
 	KubernetesInCluster  bool
 	KubernetesAPIBaseURL string
 	Namespace            string
 	LabelSelectorKey     string
 	LabelSelectorValue   string
-	SwarmPort            int
+}
+
+// Options for swarm objects.
+type Options struct {
+	swarm swarmType
+	// leaky, expected to be buffered, or errors are lost
+	Errors            chan<- error // TODO(sszuecs): should probably be hidden as implemetnation detail
+	MaxMessageBuffer  int
+	LeaveTimeout      time.Duration
+	SwarmPort         int
+	KubernetesOptions *KubernetesOptions
 }
 
 // Swarm is the main type for exchanging low latency, weakly
@@ -159,23 +88,32 @@ type Swarm struct {
 	mlist    *memberlist.Memberlist
 }
 
-func NewSwarm(kubernetesInCluster bool, kubernetesURL string) (*Swarm, error) {
-	u, err := buildAPIURL(kubernetesInCluster, kubernetesURL)
+func NewSwarm(o Options) (*Swarm, error) {
+	switch getSwarmType(o) {
+	case swarmKubernetes:
+		return newKubernetesSwarm(o)
+	default:
+		return nil, ErrUnknownSwarm
+	}
+}
+
+func newKubernetesSwarm(o Options) (*Swarm, error) {
+	u, err := buildAPIURL(o.KubernetesOptions.KubernetesInCluster, o.KubernetesOptions.KubernetesAPIBaseURL)
 	if err != nil {
-		log.Fatalf("Failed to build kubernetes API url from url %s running in cluster %v: %v", kubernetesURL, kubernetesInCluster, err)
+		return nil, fmt.Errorf("failed to build kubernetes API url from url %s running in cluster %v: %v", o.KubernetesOptions.KubernetesAPIBaseURL, o.KubernetesOptions.KubernetesInCluster, err)
 	}
 
-	o := Options{
-		Errors:               make(chan<- error), // FIXME - do WE have to read this, or...?
-		MaxMessageBuffer:     DefaultMaxMessageBuffer,
-		LeaveTimeout:         DefaultLeaveTimeout,
-		KubernetesInCluster:  kubernetesInCluster,
-		KubernetesAPIBaseURL: u,
-		Namespace:            DefaultNamespace,
-		LabelSelectorKey:     DefaultLabelSelctorKey,
-		LabelSelectorValue:   DefaultLabelSelctorValue,
-		SwarmPort:            DefaultSwarmPort,
-	}
+	o.Errors = make(chan<- error) // FIXME - do WE have to read this, or...?
+	o.KubernetesOptions.KubernetesAPIBaseURL = u
+
+	// TODO(sszuecs): use args to set defaults and use these defaults in opts parsing
+	o.MaxMessageBuffer = DefaultMaxMessageBuffer
+	o.LeaveTimeout = DefaultLeaveTimeout
+	o.SwarmPort = DefaultSwarmPort
+	o.KubernetesOptions.Namespace = DefaultNamespace
+	o.KubernetesOptions.LabelSelectorKey = DefaultLabelSelctorKey
+	o.KubernetesOptions.LabelSelectorValue = DefaultLabelSelctorValue
+
 	return Start(o)
 }
 
@@ -187,6 +125,7 @@ func Start(o Options) (*Swarm, error) {
 
 // Join will join given Swarm peers and return an initialiazed Swarm
 // object if successful.
+// TODO(sszuecs): check the options elsewhere
 func Join(o Options, self *NodeInfo, nodes []*NodeInfo) (*Swarm, error) {
 	log.Infof("SWARM: Going to join swarm of %d nodes, self=%s", len(nodes), self)
 	c := memberlist.DefaultLocalConfig()
@@ -262,7 +201,9 @@ func Join(o Options, self *NodeInfo, nodes []*NodeInfo) (*Swarm, error) {
 		shared:           shared,
 	}
 
+	// TODO(sszuecs): maybe we should wrap it in a recover for panic, but we need to close the channels
 	go s.control()
+
 	return s, nil
 }
 
@@ -385,141 +326,4 @@ func (s *Swarm) Listen(key string, c chan<- *Message) {}
 
 func (s *Swarm) Leave() {
 	close(s.leave)
-}
-
-type messageType int
-
-const (
-	sharedValue messageType = iota
-	broadcast
-)
-
-type message struct {
-	Type   messageType
-	Source string
-	Key    string
-	Value  interface{}
-}
-
-type outgoingMessage struct {
-	message *message
-	encoded []byte
-}
-
-type Message struct {
-	Source string
-	Value  interface{}
-}
-
-type reqOutgoing struct {
-	overhead int
-	limit    int
-	ret      chan [][]byte
-}
-
-// mlDelegate is a memberlist delegate
-type mlDelegate struct {
-	meta     []byte
-	outgoing chan<- reqOutgoing
-	incoming chan<- []byte
-}
-
-type sharedValues map[string]map[string]interface{}
-
-type valueReq struct {
-	key string
-	ret chan map[string]interface{}
-}
-
-// NodeMeta implements a memberlist delegate
-func (d *mlDelegate) NodeMeta(limit int) []byte {
-	if len(d.meta) > limit {
-		// TODO: would nil better here?
-		// documentation is unclear
-		return d.meta[:limit]
-	}
-
-	return d.meta
-}
-
-// NotifyMsg implements a memberlist delegate
-func (d *mlDelegate) NotifyMsg(m []byte) {
-	d.incoming <- m
-}
-
-// GetBroadcasts implements a memberlist delegate
-// TODO: verify over TCP-only
-func (d *mlDelegate) GetBroadcasts(overhead, limit int) [][]byte {
-	req := reqOutgoing{
-		overhead: overhead,
-		limit:    limit,
-		ret:      make(chan [][]byte),
-	}
-	d.outgoing <- req
-	return <-req.ret
-}
-
-// LocalState implements a memberlist delegate
-func (d *mlDelegate) LocalState(bool) []byte { return nil }
-
-// MergeRemoteState implements a memberlist delegate
-func (d *mlDelegate) MergeRemoteState(buf []byte, join bool) {}
-
-// the top level map is used internally, we can use it as mutable
-// the leaf maps are shared, we need to clone those
-func (sv sharedValues) set(source, key string, value interface{}) {
-	prev := sv[key]
-	sv[key] = make(map[string]interface{})
-	for s, v := range prev {
-		sv[key][s] = v
-	}
-
-	sv[key][source] = value
-}
-
-func reverse(b [][]byte) [][]byte {
-	for i := range b[:len(b)/2] {
-		b[i], b[len(b)-1-i] = b[len(b)-1-i], b[i]
-	}
-
-	return b
-}
-
-func takeMaxLatest(b [][]byte, overhead, max int) [][]byte {
-	var (
-		bb   [][]byte
-		size int
-	)
-
-	for i := range b {
-		bli := b[len(b)-i-1]
-
-		if size+len(bli)+overhead > max {
-			break
-		}
-
-		bb = append(bb, bli)
-		size += len(bli) + overhead
-	}
-
-	return reverse(bb)
-}
-
-func encodeMessage(m *message) ([]byte, error) {
-	// we're not saving the encoder together with the connections, because
-	// even if the reflection info would be cached, it's very fragile and
-	// complicated. These messages should be small, it should be OK to pay
-	// this cost.
-
-	var buf bytes.Buffer
-	enc := gob.NewEncoder(&buf)
-	err := enc.Encode(m)
-	return buf.Bytes(), err
-}
-
-func decodeMessage(b []byte) (*message, error) {
-	var m message
-	dec := gob.NewDecoder(bytes.NewBuffer(b))
-	err := dec.Decode(&m)
-	return &m, err
 }
